pyproject.toml
[project]
name = "assistme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "typer>=0.20.0",
    "playwright>=1.37.0",
    "python-dotenv",
]
